refactor alerting-gen to separate generate and upload from cli options
